requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
azure-common==1.1.28
azure-core==1.29.4
azure-identity==1.14.0
azure-keyvault-secrets==4.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.3
dataclasses-json==0.5.14
defusedxml==0.7.1
Django==4.2.4
django-allauth==0.55.2
django-appconf==1.0.5
django-colorfield==0.10.1
django-compressor==4.4
django-environ==0.11.1
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
isodate==0.6.1
langchain==0.0.268
langsmith==0.0.25
marshmallow==3.20.1
msal==1.24.0
msal-extensions==1.0.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
oauthlib==3.2.2
openai==0.27.8
packaging==23.1
Pillow==10.0.0
portalocker==2.8.2
psycopg2==2.9.7
pycparser==2.21
pydantic==2.2.1
pydantic_core==2.6.1
PyJWT==2.8.0
python3-openid==3.2.0
PyYAML==6.0.1
rcssmin==1.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rjsmin==1.2.1
six==1.16.0
SQLAlchemy==2.0.20
sqlparse==0.4.4
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
whitenoise==6.5.0
yarl==1.9.2
